fix: Allow to use an optional app_id argument in delete command (#1939)
* fix: Allow to use an optional app_id argument in delete command

* Keep the cloud config file if the app id differs

* Double check intention

+++ dozer-cli/src/simple/cloud_orchestrator.rs
@@ -136,10 +136,37 @@ impl CloudOrchestrator for SimpleOrchestrator {
     }
 
     fn delete(&mut self, cloud: Cloud) -> Result<(), OrchestrationError> {
-        let app_id = cloud
-            .app_id
-            .clone()
-            .unwrap_or(CloudAppContext::get_app_id(self.config.cloud.as_ref())?);
+        // Get app_id from command line argument if there, otherwise take it from the cloud config file
+        // if the app_id is from the cloud config file then set `delete_cloud_file` to true and use it later
+        // to delete the file after deleting the app
+
+        let (app_id, delete_cloud_file) = if let Some(app_id) = cloud.app_id.clone() {
+            // if the app_id on command line is equal to the one in the cloud config file then file can be deleted
+            if app_id == CloudAppContext::get_app_id(self.config.cloud.as_ref())? {
+                (app_id, true)
+            } else {
+                (app_id, false)
+            }
+        } else {
+            (
+                CloudAppContext::get_app_id(self.config.cloud.as_ref())?,
+                true,
+            )
+        };
+
+        let mut double_check = String::new();
+        println!("Are you sure to delete the application {}? (y/N)", app_id);
+        io::stdin()
+            .read_line(&mut double_check)
+            .map_err(FailedToReadOrganisationName)?;
+        let response = double_check.trim().to_string().to_uppercase();
+
+        if response == "Y" {
+            info!("Deleting application {}", app_id);
+        } else {
+            info!("The application {} was not deleted", app_id);
+            return Ok(());
+        }
 
         let cloud_config = self.config.cloud.as_ref();
         self.runtime.block_on(async move {
@@ -163,7 +190,9 @@ impl CloudOrchestrator for SimpleOrchestrator {
             if delete_result.success {
                 steps.complete_step(Some(&format!("Deleted {}", &app_id)));
 
-                let _ = CloudAppContext::delete_config_file();
+                if delete_cloud_file {
+                    let _ = CloudAppContext::delete_config_file();
+                }
             }
 
             Ok::<(), CloudError>(())
